Move extractDependencies to its own module and re-export up
