+++ src/mirage/resolvers/union.ts
@@ -1,7 +1,7 @@
 import { GraphQLSchema, GraphQLTypeResolver, GraphQLUnionType, GraphQLAbstractType } from 'graphql';
-import { extractDependencies } from '../../utils';
 import { MirageGraphQLMapper } from '../mapper';
 import { findMostInCommon, modelNameToTypeName } from './helpers';
+import { extractDependencies } from '../../resolver-map/extract-dependencies';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const mirageUnionResolver: GraphQLTypeResolver<any, any> = function (
