+++ src/resolver-map/index.ts
@@ -1,3 +1,4 @@
 export { pack } from './pack';
 export { patchEachField } from './patch-each-field';
 export { wrapEachField } from './wrap-each-field';
+export { extractDependencies } from './extract-dependencies';
